style: use uint256 in registry coordinator

+++ src/BLSRegistryCoordinatorWithIndices.sol
@@ -617,8 +617,8 @@ contract BLSRegistryCoordinatorWithIndices is EIP712, Initializable, IBLSRegistr
     ) external view returns (uint32[] memory) {
         uint32[] memory indices = new uint32[](operatorIds.length);
         for (uint256 i = 0; i < operatorIds.length; i++) {
-            uint length = _operatorIdToQuorumBitmapHistory[operatorIds[i]].length;
-            for (uint j = 0; j < length; j++) {
+            uint256 length = _operatorIdToQuorumBitmapHistory[operatorIds[i]].length;
+            for (uint256 j = 0; j < length; j++) {
                 if (_operatorIdToQuorumBitmapHistory[operatorIds[i]][length - j - 1].updateBlockNumber <= blockNumber) {
                     uint32 nextUpdateBlockNumber = 
                         _operatorIdToQuorumBitmapHistory[operatorIds[i]][length - j - 1].nextUpdateBlockNumber;
